requirements.txt
flask
yfinance
gunicorn
plotly
newsapi-python